requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
langchain-core==1.2.19
langchain-text-splitters==1.1.1
langchain-huggingface==1.2.1
huggingface_hub==1.7.1
pypdf==6.8.0
python-multipart==0.0.22
pytest==6.8.0
python-dotenv==1.2.2